@@ -4,10 +4,10 @@
   "homepage": "https://avalos010.github.io/photography-search/",
   "private": true,
   "dependencies": {
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "3.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.4",
+    "react-scripts": "^5.0.1",
     "react-spinners": "^0.5.4",
     "unsplash-js": "^5.0.0"
   },
@@ -35,9 +35,6 @@
     ]
   },
   "devDependencies": {
-    "eslint-config-prettier": "^4.3.0",
-    "eslint-plugin-prettier": "^3.1.0",
-    "gh-pages": "^2.0.1",
-    "prettier": "^1.17.1"
+    "gh-pages": "^2.0.1"
   }
 }
