@@ -1,17 +1,12 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import App from "./components/App";
-import * as serviceWorker from "./serviceWorker";
-import { HashRouter } from "react-router-dom";
+import { BrowserRouter } from "react-router-dom";
 
-ReactDOM.render(
-  <HashRouter>
+const container = document.getElementById("root");
+const root = createRoot(container); // createRoot(container!) if you use TypeScript
+root.render(
+  <BrowserRouter>
     <App />
-  </HashRouter>,
-  document.getElementById("root")
+  </BrowserRouter>
 );
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
